Dynamic session

+++ src/wpipe/core.py
@@ -6,12 +6,9 @@ Please note that this module is private. All functions and objects
 are available in the main ``wpipe`` namespace - use that instead.
 """
 import importlib
-import contextlib
 import os
 import sys
 import types
-import itertools
-import numbers
 import datetime
 import time
 import subprocess
@@ -30,13 +27,12 @@ from . import sqlintf as si
 
 __all__ = ['importlib', 'os', 'sys', 'types', 'datetime', 'time', 'subprocess',
            'logging', 'glob', 'shutil', 'warnings', 'json', 'ast', 'atexit',
-           'np', 'pd', 'si', 'PARSER', 'as_int', 'try_scalar', 'clean_path',
-           'split_path', 'remove_path', 'key_wpipe_separator',
-           'initialize_args', 'wpipe_to_sqlintf_connection',
-           'return_dict_of_attrs', 'to_json',
-           'BaseProxy', 'ChildrenProxy', 'DictLikeChildrenProxy']
+           'np', 'pd', 'si', 'PARSER', 'as_int',
+           'clean_path', 'split_path', 'remove_path', 'key_wpipe_separator',
+           'initialize_args', 'wpipe_to_sqlintf_connection', 'in_session',
+           'return_dict_of_attrs', 'to_json']
 
-PARSER = si.PARSER  # argparse.ArgumentParser()
+PARSER = si.PARSER
 PARSER.add_argument('--user', '-u', dest='user_name', type=str,
                     default=os.environ['WPIPE_USER'] if 'WPIPE_USER' in os.environ.keys()
                     else [warnings.warn("Set environment variable $WPIPE_USER to associate a default username"),
@@ -46,9 +42,6 @@ PARSER.add_argument('--pipeline', '-p', dest='pipeline', type=str, default=os.ge
                     help='Path or ID of pipeline - default to current working directory')
 PARSER.add_argument('--job', '-j', dest='job_id', type=int, help='ID of this job')
 PARSER.add_argument('--event', '-e', dest='event_id', type=int, help='ID of the firing event of this job')
-
-# if os.getcwd() not in map(os.path.abspath, sys.path):
-#     sys.path.insert(0, os.getcwd())
 
 pd.set_option('io.hdf.default_format', 'table')
 
@@ -70,26 +63,6 @@ def as_int(string):
     try:
         return int(string)
     except (ValueError, TypeError):
-        return string
-
-
-def try_scalar(string):
-    """
-    Returns given string parameter as scalar value, or return as string.
-
-    Parameters
-    ----------
-    string : string
-        Input string parameter.
-
-    Returns
-    -------
-    value
-        Scalar conversion from string or string itself.
-    """
-    try:
-        return ast.literal_eval(string)
-    except (ValueError, NameError, SyntaxError):
         return string
 
 
@@ -138,28 +111,34 @@ def split_path(path):
     return base, name, ext
 
 
-def remove_path(path, hard=False):
+def remove_path(*paths, hard=False):
     """
     Remove file or directory located at path.
 
     Parameters
     ----------
-    path : string
-        Path where file or directory is located
+    paths : string or array_like of strings
+        Paths where file or directories are located
     hard : boolean
         Flag to proceed to a hard tree removal of a directory - defaults to
         False.
     """
-    if os.path.isfile(path):
-        os.remove(path)
-    elif os.path.isdir(path):
-        if hard:
-            shutil.rmtree(path)
-        else:
-            try:
-                os.rmdir(path)
-            except OSError:
-                warnings.warn("Cannot remove directory %s : directory is not empty." % path)
+    if len(paths):
+        if len(paths) == 1 and hasattr(paths[0], '__len__') and not isinstance(paths[0], str):
+            paths = paths[0]
+        for path in paths:
+            if os.path.isfile(path):
+                os.remove(path)
+            elif os.path.isdir(path):
+                if hard:
+                    shutil.rmtree(path)
+                else:
+                    try:
+                        os.rmdir(path)
+                    except OSError:
+                        warnings.warn("Cannot remove directory %s : directory is not empty." % path)
+    else:
+        raise TypeError('remove_path expected at least 1 arguments, get 0')
 
 
 def key_wpipe_separator(obj):
@@ -228,6 +207,45 @@ def wpipe_to_sqlintf_connection(cls, cls_name):
         cls._inst = super(getattr(sys.modules['wpipe'], cls_name), cls).__new__(cls)
         getattr(cls, cls_attr)._wpipe_object = cls._inst
         setattr(cls._inst, cls_attr, getattr(cls, cls_attr))
+        setattr(cls._inst, '_session', None)
+
+
+def in_session(si_attr, **local_kw):
+    """
+    Returns a decorator that places the modified function in a begin_session
+    context manager statement. The modified must be defined around a first
+    positional argument which object will hold the associated BeginSession
+    object as attribute '_session'. Also, it adds to the corresponding Session
+    the object that is hold in the attribute designated by si_attr of that
+    same object in first positional argument of the modified function.
+
+    Parameters
+    ----------
+    si_attr : string
+        Name of the attribute where to find the object to add to the Session.
+
+    local_kw : dict
+        TODO
+
+    Returns
+    -------
+    decor :
+        Decorator to be used.
+    """
+
+    def decor(func):
+        def wrapper(self_cls, *args, **kwargs):
+            with si.begin_session(**local_kw) as session:
+                _temp = self_cls._session
+                self_cls._session = session
+                session.add(getattr(self_cls, si_attr))
+                output = func(self_cls, *args, **kwargs)
+                self_cls._session = _temp
+                return output
+
+        return wrapper
+
+    return decor
 
 
 def return_dict_of_attrs(obj):
@@ -245,12 +263,14 @@ def return_dict_of_attrs(obj):
     attrs : dict
         Dictionary of attributes of obj.
     """
-    si.commit()
-    return dict((attr, getattr(obj, attr))
-                for attr in dir(obj) if attr[0] != '_'
-                and getattr(obj, attr) is not None
-                and type(getattr(obj, attr)).__module__.split('.')[0]
-                not in ['sqlalchemy', 'wpipe'])
+    with si.begin_session() as session:
+        session.add(obj)
+        session.refresh(obj)
+        return dict((attr, getattr(obj, attr))
+                    for attr in dir(obj) if attr[0] != '_'
+                    and getattr(obj, attr) is not None
+                    and type(getattr(obj, attr)).__module__.split('.')[0]
+                    not in ['sqlalchemy', 'wpipe'])
 
 
 def to_json(obj, *args, **kwargs):
@@ -266,254 +286,3 @@ def to_json(obj, *args, **kwargs):
     """
     pd.DataFrame(return_dict_of_attrs(obj),
                  index=[0]).to_json(*args, **kwargs)
-
-
-class BaseProxy:
-    def __new__(cls, *args, **kwargs):
-        if cls is BaseProxy:
-            proxy = getattr(kwargs.pop('parent', None),
-                            kwargs.pop('attr_name', ''))
-            if kwargs.pop('try_scalar', False):
-                proxy = try_scalar(proxy)
-            if isinstance(proxy, str):
-                cls = StringProxy
-            elif isinstance(proxy, numbers.Number):
-                cls = NumberProxy
-            elif isinstance(proxy, datetime.datetime):
-                cls = DatetimeProxy
-            else:
-                raise ValueError("Invalid proxy type %s" % type(proxy))
-            args = [proxy]
-            return cls.__new__(cls, *args, *kwargs)
-        return super().__new__(cls, *args, *kwargs)
-
-    def __init__(self, *args, **kwargs):
-        self._parent = kwargs.pop('parent', None)
-        self._parent_id = int(self.parent.id)
-        self._attr_name = kwargs.pop('attr_name', '')
-        self._try_scalar = kwargs.pop('try_scalar', False)
-
-    @property
-    def parent(self):
-        return self._parent
-
-    @property
-    def parent_id(self):
-        return self._parent_id
-
-    @property
-    def attr_name(self):
-        return self._attr_name
-
-    @property
-    def try_scalar(self):
-        return self._try_scalar
-
-
-class StringProxy(BaseProxy, str):
-    pass
-
-
-class NumberProxy(BaseProxy):
-    def __new__(cls, *args, **kwargs):
-        if cls is NumberProxy:
-            proxy = args[0]
-            if isinstance(proxy, int):
-                cls = IntProxy
-            elif isinstance(proxy, float):
-                cls = FloatProxy
-            else:
-                raise ValueError("Invalid proxy type %s" % type(proxy))
-            return cls.__new__(cls, *args, *kwargs)
-        return super().__new__(cls, *args, *kwargs)
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(self, *args, **kwargs)
-
-    def __iadd__(self, other):
-        for retry in si.retrying_nested():
-            with retry:
-                _temp = si.session.query(self.parent.__class__).with_for_update().filter_by(id=self.parent_id).one()
-                setattr(_temp, self.attr_name,
-                        [lambda x:x, try_scalar][self._try_scalar](getattr(_temp, self.attr_name)) + other)
-                _temp = BaseProxy(parent=self.parent,
-                                  attr_name=self.attr_name,
-                                  try_scalar=self.try_scalar)
-                retry.retry_state.commit()
-        return _temp
-
-
-class IntProxy(NumberProxy, int):
-    pass
-
-
-class FloatProxy(NumberProxy, float):
-    pass
-
-
-class DatetimeProxy(BaseProxy, datetime.datetime):
-    def __new__(cls, *args, **kwargs):
-        if cls is DatetimeProxy:
-            args = [
-                args[0].year,
-                args[0].month,
-                args[0].day,
-                args[0].hour,
-                args[0].minute,
-                args[0].second,
-                args[0].microsecond,
-                args[0].tzinfo
-            ]
-        return super().__new__(cls, *args, *kwargs)
-
-
-class ChildrenProxy:
-    """
-        Proxy to access children of a Wpipe object.
-
-        Parameters
-        ----------
-        parent : sqlintf object
-            Parent sqlintf object of Wpipe object associated to this proxy.
-        children_attr : string
-            Attribute of the sqlintf object returning its children.
-        cls_name : string
-            Name of the class of the children.
-        child_attr : string
-            Child attribute that distinguishes the children from one another.
-    """
-
-    def __init__(self, parent, children_attr, cls_name, child_attr='name'):
-        self._parent = parent
-        self._children_attr = children_attr
-        self._cls_name = cls_name
-        self._child_attr = child_attr
-        self._work_with_sqlintf = 0
-
-    def __repr__(self):
-        self._refresh()
-        return 'Children(' + ', '.join(
-            map(lambda child: self._cls_name + '(' + repr(getattr(child, self._child_attr)) + ')',
-                self.children)) + ')'
-
-    def __len__(self):
-        self._refresh()
-        return len(self.children)
-
-    def __iter__(self):
-        self.n = 0
-        self.len = len(self)  # CURRENTLY NEEDED FOR DELETION ISSUES
-        return self
-
-    def __next__(self):
-        self.n -= self.len - len(self)  # CURRENTLY NEEDED FOR DELETION ISSUES
-        self.len = len(self)  # CURRENTLY NEEDED FOR DELETION ISSUES
-        if 0 <= self.n < self.len:
-            result = self[self.n]
-            self.n += 1
-            return result
-        else:
-            del self.n, self.len
-            raise StopIteration
-
-    def __getitem__(self, item):
-        if np.ndim(item) == 0:
-            if self._work_with_sqlintf == 0:
-                self._refresh()
-                if hasattr(self.children[item], '_wpipe_object'):
-                    return self.children[item]._wpipe_object
-                else:
-                    return getattr(sys.modules['wpipe'], self._cls_name)(self.children[item])
-            else:
-                return self.children[item]
-        else:
-            return [self[i] for i in np.arange(len(self))[item]]
-
-    def __getattr__(self, item):
-        if hasattr(getattr(sys.modules['wpipe'], self._cls_name), item):
-            self._refresh()
-            with self._with_sqlintf():
-                return np.array([getattr(self[i], item) for i in range(len(self))])
-
-    @contextlib.contextmanager
-    def _with_sqlintf(self):
-        self._work_with_sqlintf += 1
-        try:
-            yield
-        finally:
-            self._work_with_sqlintf -= 1
-
-    @property
-    def children(self):
-        return getattr(self._parent, self._children_attr)
-
-    def _refresh(self, **kwargs):
-        if self._work_with_sqlintf == 0:
-            si.refresh(self._parent, **kwargs)
-            for child in self.children:
-                si.refresh(child, **kwargs)
-
-
-class DictLikeChildrenProxy(ChildrenProxy):
-    """
-        Proxy to access children of a Wpipe object in a dictionary format.
-
-        Parameters
-        ----------
-        parent : sqlintf object
-            Parent sqlintf object of Wpipe object associated to this proxy.
-        children_attr : string
-            Attribute of the sqlintf object returning its children.
-        cls_name : string
-            Name of the class of the children.
-        child_attr : string
-            Child attribute that distinguishes the children from one another.
-        child_value : string
-            Child attribute that returns its stored value.
-    """
-
-    def __init__(self, parent, children_attr, cls_name, child_attr='name', child_value='value'):
-        super(DictLikeChildrenProxy, self).__init__(parent, children_attr, cls_name, child_attr)
-        self._child_value = child_value
-
-    def __repr__(self):
-        self._refresh()
-        return repr(dict(self._items))
-
-    def __getitem__(self, item):
-        if isinstance(item, int):
-            return super(DictLikeChildrenProxy, self).__getitem__(item)
-        self._refresh()
-        _temp = self._keys_children
-        try:
-            key = child = None
-            while key != item:
-                key, child = next(_temp)
-            return BaseProxy(parent=child,
-                             attr_name=self._child_value,
-                             try_scalar=True)
-        except StopIteration:
-            raise KeyError(item)
-
-    def __setitem__(self, item, value):
-        if not isinstance(value, BaseProxy):
-            self._refresh()
-            _temp = self._keys_children
-            try:
-                key = None
-                count = -1
-                while key != item:
-                    key, child = next(_temp)
-                    count += 1
-                child = self.children[count]
-                setattr(child, self._child_value, value)
-            except StopIteration:
-                raise KeyError(item)
-
-    @property
-    def _keys_children(self):
-        return map(lambda child: (getattr(child, self._child_attr), child), self.children)
-
-    @property
-    def _items(self):
-        return itertools.starmap(lambda key, child: (key, getattr(child, self._child_value)), self._keys_children)
